chore(uptime): Move UptimeDomainCheckFailure to `uptime/grouptype.py`
We want to keep grouptypes in the module that is related to them, so just moving this out.
